% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sushi.R
\docType{data}
\name{sushi}
\alias{sushi}
\title{CSLS raw fish data}
\format{
A list of four data frames:
\describe{
\item{pleas}{data for Pleasant Lake}
\item{pleasALW}{age/length/weight data for Pleasant Lake}
\item{long}{data for Long Lake}
\item{longALW}{age/length/weight for Long Lake}
}
}
\usage{
data(sushi)
}
\description{
Species, dates, techniques, some age/length/weight data
}
\details{
Downloaded from the fisheries database, searching by WBIC for each lake.

Raw data is processed in with the script "format_sushi" in the data-raw folder
}
\examples{
# Load fish data for Long Lake directly
lake  <- "pleas"
sushi <- CSLSdata::sushi[[lake]]

# Load fish data, then subset for specific lake
data(sushi)
sushi <- sushi$pleas

}
\keyword{datasets}
